#1079 add new test for _grid_property_reference and move GOLoop tests
